requirements.txt
asgiref==3.5.0
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
